requirements.txt
pystray==0.19.5
Pillow>=9.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
